refactor: complete Phase 5 - remove monolithic supabaseService
- Update UI components to use modular services instead of supabaseService imports
- Move SupabaseError and UserProfile types to shared types file
- Update all service interfaces and implementations to import from shared types
- Delete monolithic supabaseService.ts file
- Export service instances from RepositoryFactory for UI component access

BREAKING CHANGE: supabaseService.ts removed, all functionality now in modular services
