+++ src/services/conversation/IConversationService.ts
@@ -1,7 +1,7 @@
 import type { Conversation } from "@/store/conversationStore";
 // Import Message from journalStore as it has all the properties the DB supports
 import type { Message } from "@/store/journalStore";
-import type { SupabaseError } from "@/services/supabaseService";
+import type { SupabaseError } from "@/types/shared";
 
 /**
  * Interface for conversation/thread service
