+++ src/services/journal/IJournalService.ts
@@ -1,5 +1,5 @@
 import type { JournalEntry } from "@/store/journalStore";
-import type { SupabaseError } from "@/services/supabaseService";
+import type { SupabaseError } from "@/types/shared";
 
 /**
  * Interface for journal service
